ensure event listener is activated (hooks)
The OS X delegate listener for an event is only activated if the
`on` call is made for the event id.
This is an optimization to prevent flooding the application with
events that the application does not actually care about.
The `impl.on` call is what activates the response handler for the
event in question.  This logic was missing in the RegisterHooks()
function but is present in the `On()` function.  In the case of using
`On()` first followed by `RegisterHooks()` things work correctly.